+++ tetrs/src/pt.rs
@@ -7,12 +7,12 @@ use ::std::ops;
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 #[repr(C)]
 pub struct Point {
-	pub x: i32,
-	pub y: i32,
+	pub x: i8,
+	pub y: i8,
 }
 
 impl Point {
-	pub fn new(x: i32, y: i32) -> Point {
+	pub fn new(x: i8, y: i8) -> Point {
 		Point {
 			x: x,
 			y: y,
